refactor: restore runtime package quality gates
