+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/dispatch_worker_sessions_cutover.go
@@ -11,7 +11,6 @@ import (
 
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
-	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/state"
 	providersessions "github.com/portpowered/infinite-you/pkg/services/provider_sessions"
 	"github.com/portpowered/infinite-you/pkg/services/providers"
 	"github.com/portpowered/infinite-you/pkg/services/recordings"
@@ -19,151 +18,6 @@ import (
 	"github.com/portpowered/infinite-you/pkg/services/workers"
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 )
-
-// WorkstationRequestExecutorConfig supplies the immutable Runtime facts needed
-// to resolve a direct Worker Session request before handing it to the shared
-// Workers Execute boundary. It is deliberately separate from the Runtime's
-// active attempt lifecycle: direct Worker Sessions own their own supervision.
-type WorkstationRequestExecutorConfig struct {
-	Service                    workers.Service
-	RuntimeDefinitions         interfaces.RuntimeDefinitionLookup
-	InvocationInterpolation    interfaces.InvocationInterpolationService
-	InvocationFileReader       interfaces.FileReader
-	WorkflowContext            *workers.Context
-	FactorySessionID           string
-	RuntimeID                  string
-	RecordingID                string
-	EventHistory               recordings.RuntimeLedger
-	NewID                      factory.IDGenerator
-	PromptRenderer             runtimePromptRenderer
-	TemplateFieldResolver      runtimeTemplateFieldResolver
-	PromptSourceReader         interfaces.FileReader
-	MockWorkers                *workers.MockWorkersConfig
-	ProgressPublisher          workers.ProgressPublisher
-	Net                        *state.Net
-	ExpectedArtifactFileSystem any
-}
-
-type PromptRenderer = runtimePromptRenderer
-type TemplateFieldResolver = runtimeTemplateFieldResolver
-
-// NewWorkstationRequestExecutor creates the Runtime-owned compatibility
-// adapter used by top-level Worker Session routes. It resolves minimal legacy
-// direct requests from the immutable Factory definition and then invokes the
-// process-scoped Workers service with a complete detached ExecuteRequest.
-func NewWorkstationRequestExecutor(
-	config WorkstationRequestExecutorConfig,
-) workers.WorkstationRequestExecutor {
-	if config.Service == nil {
-		return nil
-	}
-	return &workstationRequestExecutor{
-		service: config.Service,
-		cfg: &runtimeConfig{
-			executeService:             config.Service,
-			runtimeConfig:              config.RuntimeDefinitions,
-			invocationInterpolation:    config.InvocationInterpolation,
-			invocationFileReader:       config.InvocationFileReader,
-			workflowContext:            config.WorkflowContext.Clone(),
-			recordingID:                strings.TrimSpace(config.RecordingID),
-			runtimeID:                  strings.TrimSpace(config.RuntimeID),
-			eventHistory:               config.EventHistory,
-			newID:                      config.NewID,
-			promptRenderer:             config.PromptRenderer,
-			templateFieldResolver:      config.TemplateFieldResolver,
-			promptSourceReader:         config.PromptSourceReader,
-			mockWorkersConfig:          config.MockWorkers.Clone(),
-			progressPublisher:          config.ProgressPublisher,
-			net:                        config.Net,
-			expectedArtifactFileSystem: expectedArtifactFileSystemFrom(config.ExpectedArtifactFileSystem),
-		},
-		sessionID: strings.TrimSpace(config.FactorySessionID),
-	}
-}
-
-type workstationRequestExecutor struct {
-	service   executeCapability
-	cfg       *runtimeConfig
-	sessionID string
-}
-
-// SetRuntimeLogger binds the opened Runtime's log sink to this compatibility
-// adapter. The adapter itself remains detached; only the per-attempt logger
-// capability crosses into the process-scoped Workers request.
-func (executor *workstationRequestExecutor) SetRuntimeLogger(logger factory.Logger) {
-	if executor == nil || executor.cfg == nil {
-		return
-	}
-	executor.cfg.logger = logger
-}
-
-func isRuntimePromptRenderError(err error) bool {
-	var promptErr *runtimePromptRenderError
-	return errors.As(err, &promptErr)
-}
-
-func (executor workstationRequestExecutor) Execute(
-	ctx context.Context,
-	request workers.WorkstationExecutionRequest,
-) (workers.WorkResult, error) {
-	if executor.service == nil || executor.cfg == nil {
-		return workers.WorkResult{}, workers.ErrExecuteUnavailable
-	}
-	request = workers.CloneWorkstationExecutionRequest(request)
-	if strings.TrimSpace(request.FactorySessionID) == "" {
-		request.FactorySessionID = executor.sessionID
-	}
-	request.WorkerType = firstRuntimeValue(request.WorkerType, request.Dispatch.WorkerType)
-	if strings.TrimSpace(request.WorkstationType) == "" {
-		request.WorkstationType = firstRuntimeValue(request.WorkstationType, request.Dispatch.WorkstationName)
-	}
-	request = runtimeRecordingExecutionRequest(executor.cfg, request)
-	workstationRequest := workers.WorkstationDispatchRequest{
-		WorkstationName: firstRuntimeValue(request.Dispatch.WorkstationName, request.WorkstationType),
-		Execution:       request,
-	}
-	executeRequest, err := executeRequestFromWorkstationRequest(executor.cfg, workstationRequest)
-	if err != nil {
-		result := workers.WorkResult{
-			DispatchID:   request.Dispatch.DispatchID,
-			TransitionID: request.Dispatch.TransitionID,
-			Outcome:      workers.OutcomeFailed,
-			Error:        err.Error(),
-		}
-		if isRuntimePromptRenderError(err) {
-			result.Error = "prompt render failed: " + err.Error()
-			return result, nil
-		}
-		return result, err
-	}
-	result, executeErr := executor.service.Execute(ctx, executeRequest)
-	result = normalizeDetachedExecutionResult(executor.cfg, executeRequest, result)
-	dispatchResult, dispatchErr := workstationDispatchResultFromExecute(
-		workstationDispatchRequestForResult(workstationRequest, executeRequest),
-		result,
-		executeErr,
-	)
-	return dispatchResult.Result, dispatchErr
-}
-
-func runtimeRecordingExecutionRequest(
-	cfg *runtimeConfig,
-	request workers.WorkstationExecutionRequest,
-) workers.WorkstationExecutionRequest {
-	if cfg == nil {
-		return request
-	}
-	if strings.TrimSpace(request.RuntimeID) == "" {
-		request.RuntimeID = strings.TrimSpace(cfg.runtimeID)
-	}
-	if strings.TrimSpace(request.RecordingID) == "" {
-		request.RecordingID = strings.TrimSpace(cfg.recordingID)
-	}
-	if strings.TrimSpace(request.GenerationID) == "" && cfg.eventHistory != nil {
-		request.GenerationID = strings.TrimSpace(cfg.eventHistory.StreamGenerationID())
-	}
-	return request
-}
 
 // startThroughWorkerSessions is the W4 Runtime dispatch cutover seam. For
 // every resolved dispatch it: boots the underlying Workers pool if needed,
